projects/crew-ai-engineering-team/backend/pyproject.toml
[project]
name = "engineering-team"
version = "0.1.0"
description = "AI Engineering Team powered by CrewAI - 5 agents collaborating to build production-ready apps"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pydantic",
    "sse-starlette",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
